refactor: extract prompt templates from scheduler to dedicated module
Move prompt generation functions to dedicated prompt-templates.ts module:
- generateSeniorPrompt
- generateIntermediatePrompt
- generateJuniorPrompt
- generateQAPrompt

Benefits:
- Separates concerns: scheduler focuses on orchestration, prompts are in dedicated module
- Easier to maintain and update agent prompts
- Cleaner scheduler.ts file (~350 lines reduced)
- More testable: prompts can be tested independently
- Facilitates prompt version management in future

All tests passing (379 tests).
